Large overhaul
this commit makes SubPath the primary abstraction of this package

There is still support for the basic svg drawing instructions, but
ideally that is not the level of abstraction you work in.

Rather, curve is for drawing (list of points -> some line),
and SubPath is the composition abstraction, with support for connect
two subpaths with a straight line, continue (pasting path b at the end
of path a) and continue smooth (pasting b at the end of a and rotating b
such that they fit together smoothly - the derivatives are equal).

Then, a path is only a layering system (a on top of b)